cloud_run_service/main/requirements.txt
flask==3.0.3
google-cloud-pubsub==2.23.0
google-cloud-firestore==2.18.0
google-cloud-logging==3.11.2
python-dateutil==2.9.0.post0